move decorators to testing/utils.py (#1761)
* move decorators to testing/utils.py

* add import

* fix import

* fix ruff formatting error

* ruff fixes

* ruff format

* compute_capability test

* update

* update rest of tests

* fix ruff
