+++ test/float8/test_base.py
@@ -14,11 +14,11 @@ import pytest
 import torch
 import torch.nn as nn
 
+from torchao.testing.utils import skip_if_rocm
 from torchao.utils import (
     TORCH_VERSION_AT_LEAST_2_5,
     is_sm_at_least_89,
     is_sm_at_least_90,
-    skip_if_rocm,
 )
 
 if not TORCH_VERSION_AT_LEAST_2_5:
