+++ test/hqq/test_hqq_affine.py
@@ -9,9 +9,9 @@ from torchao.quantization import (
     quantize_,
     uintx_weight_only,
 )
+from torchao.testing.utils import skip_if_rocm
 from torchao.utils import (
     TORCH_VERSION_AT_LEAST_2_3,
-    skip_if_rocm,
 )
 
 cuda_available = torch.cuda.is_available()
