+++ test/prototype/test_awq.py
@@ -5,10 +5,10 @@ import pytest
 import torch
 
 from torchao.quantization import quantize_
+from torchao.testing.utils import skip_if_rocm
 from torchao.utils import (
     TORCH_VERSION_AT_LEAST_2_3,
     TORCH_VERSION_AT_LEAST_2_5,
-    skip_if_rocm,
 )
 
 if TORCH_VERSION_AT_LEAST_2_3:
