+++ test/prototype/test_splitk.py
@@ -14,7 +14,7 @@ except ImportError:
     triton_available = False
 
 
-from torchao.utils import skip_if_compute_capability_less_than, skip_if_rocm
+from torchao.testing.utils import skip_if_compute_capability_less_than, skip_if_rocm
 
 
 @unittest.skipIf(not triton_available, "Triton is required but not available")
