+++ test/quantization/test_quant_api.py
@@ -48,6 +48,7 @@ from torchao.quantization.subclass import (
     Int8WeightOnlyQuantizedLinearWeight,
 )
 from torchao.quantization.utils import compute_error
+from torchao.testing.utils import skip_if_rocm
 from torchao.utils import (
     TORCH_VERSION_AT_LEAST_2_3,
     TORCH_VERSION_AT_LEAST_2_4,
@@ -55,7 +56,6 @@ from torchao.utils import (
     TORCH_VERSION_AT_LEAST_2_6,
     is_sm_at_least_89,
     is_sm_at_least_90,
-    skip_if_rocm,
     unwrap_tensor_subclass,
 )
